Refactor token validity and status check from main SDK, refactor static stuff into ApprienUtility class
