+++ src/ApprienUnitySDK/Assets/Apprien/Scripts/ApprienProduct.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 50da5c0c8fac31b42b7a4e3bc525fc93
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
