+++ src/ApprienUnitySDK/Assets/Apprien/Scripts/ApprienUtility.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7e857ff19243a9945a66c75c12594453
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
